+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
 ## [0.0.8] - 2026-Apr-13
 - basic structure for nfty → NFTables config generation
 - `default-rules` generation logic implemented (eg `established,related`, `drop: invalid`, etc)
 - dport formatting function, such as for accepting port-ranges, port lists, etc.
 - srcip formatting function, such as for accepint address-list names OR arrays of individual IPs
 - ipv6 functionality for all of the above
+- additional nfty → NFTables config generation
+- chain, socket, and src-ip handling
+- interface matching generation
 
 ## [0.0.7] - 2026-Apr-12
 - the name is bond, james bond